[js] [jscoq] Use ES6 modules step 8: jscoq object
